+++ reapy/core/item/take.py
@@ -16,7 +16,20 @@ class Take(ReapyObject):
 
     @property
     def _args(self):
-        return (self.id,)
+        return self.id,
+
+    def add_audio_accessor(self):
+        """
+        Create audio accessor and return it.
+
+        Returns
+        -------
+        audio_accessor : AudioAccessor
+            Audio accessor on take.
+        """
+        audio_accessor_id = RPR.CreateTakeAudioAccessor(self.id)
+        audio_accessor = reapy.AudioAccessor(audio_accessor_id)
+        return audio_accessor
 
     def add_fx(self, name, even_if_exists=True):
         """